Project1/requirements.txt
streamlit
psycopg2-binary
pandas
google-generativeai
toml
streamlit-calendar
requests
plotly
python-dateutil
requests
torch
torchvision
torchaudio